backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RoundZero AI Interview Coach Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.6.1",
    "httpx>=0.27.2",
    "google-genai>=1.37.0",
    "google-adk>=0.1.0",
    "groq>=0.11.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.22",
    "asyncpg>=0.31.0",
    "sse-starlette>=3.2.0",
    "upstash-redis>=1.0.0",
    "redis>=5.2.0",
    "elevenlabs>=1.50.0",
    "pinecone>=6.0.1",
    "supermemory>=1.0.14",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.10.0",
    "aiortc>=1.9.0",
    "av>=12.0.0",
    "aiolimiter>=1.1.0",
    "wsproto>=1.3.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]
vision = [
    "vision-agents[gemini,deepgram,elevenlabs,getstream]>=0.3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.6.9",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
]

[tool.ruff]
line-length = 110
target-version = "py311"
lint.select = ["E", "F", "W", "B", "I", "UP", "ASYNC"]
